+++ generateForm.js
@@ -1,0 +1,28 @@
+module.exports = function generateForm(path) {
+  const form = `<!DOCTYPE html>
+<html lang="ja">
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width,initial-scale=1">
+<title>Login form</title>
+</head>
+<body>
+<form action="${path}" method="POST">
+  <table>
+    <tbody>
+      <tr>
+        <td>ユーザ名：</td>
+        <td><input id="username" type="text" name="username" width="10"></td>
+      </tr>
+      <tr>
+        <td>パスワード：</td>
+        <td><input id="password" type="password" name="password" width="10"></td>
+      </tr>
+    </tbody>
+  </table>
+  <button type="submit">ログイン</button>
+</form>
+</body>
+</html>`;
+  return form;
+};
